tests/wpt/tests/css/css-grid/grid-lanes/subgrid/grid-subgridded-to-grid-lanes/track-sizing/column-subgrid-auto-fit-002.html

<!DOCTYPE html>
<html>
<title>CSS Grid Lanes Test: auto-fit columns collapse around an auto-placed subgrid</title>
<link rel="help" href="https://drafts.csswg.org/css-grid-3/#subgrids">
<link rel="author" title="Alison Maher" href="mailto:almaher@microsoft.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.grid-lanes {
    display: grid-lanes;
    flow-tolerance: 0;
    grid-template-columns: repeat(auto-fit, 100px);
    width: 300px;
    height: 200px;
}

.subgrid {
    display: grid;
    grid-template-columns: subgrid;
}

.subgrid > div {
    width: 100px;
    height: 100px;
}
</style>
<div class="grid-lanes">
  <div class="subgrid" style="grid-column: 1;">
    <div></div>
  </div>
  <div class="subgrid" style="grid-column: span 1;">
    <div></div>
  </div>
</div>
<script>
  test(function() {
    const container = document.querySelector('.grid-lanes');
    const computedStyle = window.getComputedStyle(container);
    assert_equals(computedStyle.getPropertyValue('grid-template-columns'),
      "100px 0px 0px");
  });
</script>
</html>
